Translator: Make vec/matrix size getters unsigned
TType stored type's primary and secondary sizes as `unsigned char`, but
the getters returned `int`.  This caused unnecessary casts when the size
was passed from one TType to another, as well as comparisons with other
unsigned numbers.

This change specifies the type of these members as `uint8_t` and makes
the getters return the same type.  The call sites are accordingly
adjusted to remove unnecessary casts, use the correct type in local
variables, and add casts when passed to ostream::operator<<.

Bug: angleproject:6755
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/3402850
Commit-Queue: Shahbaz Youssefi <syoussefi@chromium.org>

+++ src/compiler/translator/TranslatorMetalDirect/EmitMetal.cpp
@@ -961,11 +961,12 @@ void GenMetalTraverser::emitType(const TType &type, const EmitTypeConfig &etConf
 
     if (type.isVector())
     {
-        mOut << type.getNominalSize();
+        mOut << static_cast<uint32_t>(type.getNominalSize());
     }
     else if (type.isMatrix())
     {
-        mOut << type.getCols() << "x" << type.getRows();
+        mOut << static_cast<uint32_t>(type.getCols()) << "x"
+             << static_cast<uint32_t>(type.getRows());
     }
 
     if (!isUBO)
@@ -1531,10 +1532,10 @@ bool GenMetalTraverser::visitBinary(Visit, TIntermBinary *binaryNode)
                 }
                 else
                 {
-                    int maxSize;
+                    uint32_t maxSize;
                     if (leftType.isArray())
                     {
-                        maxSize = static_cast<int>(leftType.getOutermostArraySize()) - 1;
+                        maxSize = leftType.getOutermostArraySize() - 1;
                     }
                     else
                     {
